project fix.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2008 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ simpleraytracer/camera.cpp
@@ -161,8 +161,8 @@ Camera::Camera(
 		try
 		{
 			//aperture_preview_image.saveToPng(FileUtils::join(base_indigo_path, "aperture_preview.png"), metadata, 0);
-			Bitmap ldr_image(aperture_preview_image.getWidth(), aperture_preview_image.getHeight(), 3, NULL);
-			aperture_preview_image.copyRegionToBitmap(ldr_image, 0, 0, aperture_preview_image.getWidth(), aperture_preview_image.getHeight());
+			Bitmap ldr_image((unsigned int)aperture_preview_image.getWidth(), (unsigned int)aperture_preview_image.getHeight(), 3, NULL);
+			aperture_preview_image.copyRegionToBitmap(ldr_image, 0, 0, (unsigned int)aperture_preview_image.getWidth(), (unsigned int)aperture_preview_image.getHeight());
 			PNGDecoder::write(ldr_image, metadata, FileUtils::join(appdata_path, "aperture_preview.png"));
 		}
 		catch(ImageExcep& e)
